Cleanup the code generation model interface.
Speed estimate now goes out of Motion Controller.

Removed the Control Command signal.

This is to avoid weird code generation where outputs of one
subsystem appear as outputs in other subsystem even though
they are actually an input.

Example: Speed estimate is output of Foc subsystem, and it is
used by the MotnCtrl subsystem (input). When generating code,
it actually appears as output of both subsystems!

Need to keep an eye on this behavior.