Add compile switch for UART Rx interrupts.
This allows the command console to run  in a low priority task
in the RTOS context instead of running based off interrupts.
The Rx buffer is checked every 1-10ms for new rx chars.

If trying to run automated commands via python, for example, then
interrupts may be needed to not miss any characters.