First draft. Not quite there yet.
Another PID controller added at the PWM DQ subsystem level.
It only becomes active if the actual currents are above some
calibratable threshold. If not, the PID is just a pass through.
A better way to implement would be to always have current control
enable, and have it work as a limiter when in PWM mode. That way
we save us an extra couple of PID controllers.